NAS-116395: adapt to pools-dashboard changes
