+++ src/assets/styles/mixins/cards.scss
@@ -105,3 +105,10 @@
     }
   }
 }
+
+@mixin loader-block {
+  background: var(--fg2);
+  border-radius: 3px;
+  height: 20px;
+  opacity: 0.3;
+}
